hypervisor: change vcpu variable size
Modified vcpu related information to support
more vcpu's

+++ hypervisor/src/mshv/mod.rs
@@ -1709,7 +1709,7 @@ impl vm::Vm for MshvVm {
     ///
     fn create_vcpu(
         &self,
-        id: u8,
+        id: u32,
         vm_ops: Option<Arc<dyn VmOps>>,
     ) -> vm::Result<Arc<dyn cpu::Vcpu>> {
         let vcpu_fd = self
